backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
geoalchemy2==0.15.2
shapely==2.0.6
geopandas==1.0.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
APScheduler==3.10.4
google-api-python-client==2.149.0
earthengine-api==0.1.395
reportlab==4.2.2
jinja2==3.1.4

